Refactored based on comments
