+++ tests/test_docs/script_helper.py
@@ -17,4 +17,30 @@
 #
 # ------------------------------------------------------------------------------
 
-"""This module contains the tests for the code-blocks in the use-cases folder."""
+"""This module contains helper function to extract code from the .md files."""
+
+from pathlib import Path
+
+from tests.test_docs.helper import extract_code_blocks, read_md_file
+from tests.conftest import ROOT_DIR
+
+FILEPATH = Path(ROOT_DIR, "docs", "orm-integration-to-generic.md")
+DESTINATION_PATH = Path(ROOT_DIR, "tests", "test_docs", "test_bash_yaml", "md_files", "bash-orm-integration-to-generic.md")
+
+code_blocks = extract_code_blocks(file=FILEPATH, filter="bash")
+file_to_write = ""
+for blocks in code_blocks:
+    file_to_write += "``` bash \n" + blocks + "``` \n"
+
+code_blocks = extract_code_blocks(file=FILEPATH, filter="yaml")
+for blocks in code_blocks:
+    file_to_write += "``` yaml \n" + blocks + "``` \n"
+
+with open(DESTINATION_PATH, "w+") as file:
+    file.write(file_to_write)
+
+
+
+
+
+
