+++ tests/test_docs/test_bash_yaml/md_files/bash-generic-skills.md
@@ -1,34 +1,60 @@
-``` bash
+``` bash 
 python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
-```
-
-``` bash
+``` 
+``` bash 
 aea create my_seller_aea
 cd my_seller_aea
 aea add connection fetchai/oef:0.1.0
 aea add skill fetchai/generic_seller:0.1.0
 aea install
-```
-
-``` bash
+``` 
+``` bash 
 aea create my_buyer_aea
 cd my_buyer_aea
 aea add connection fetchai/oef:0.1.0
 aea add skill fetchai/generic_buyer:0.1.0
 aea install
-```
-
-```bash
+``` 
+``` bash 
 aea generate-key fetchai
 aea add-key fetchai fet_private_key.txt
-```
-
-```bash
+``` 
+``` bash 
 aea generate-key ethereum
 aea add-key ethereum eth_private_key.txt
-```
-
-``` yaml
+``` 
+``` bash 
+aea generate-wealth fetchai
+``` 
+``` bash 
+aea generate-wealth ethereum
+``` 
+``` bash 
+addr: ${OEF_ADDR: 127.0.0.1}
+``` 
+``` bash 
+aea add connection fetchai/oef:0.1.0
+aea install
+aea run --connections fetchai/oef:0.1.0
+``` 
+``` bash 
+cd ..
+aea delete my_seller_aea
+aea delete my_buyer_aea
+``` 
+``` yaml 
+ledger_apis:
+  fetchai:
+    network: testnet
+``` 
+``` yaml 
+ledger_apis:
+  ethereum:
+    address: https://ropsten.infura.io/v3/f00f7b3ba0e848ddbdc8941c527447fe
+    chain_id: 3
+    gas_price: 50
+``` 
+``` yaml 
 |----------------------------------------------------------------------|
 |         FETCHAI                   |           ETHEREUM               |
 |-----------------------------------|----------------------------------|
@@ -59,9 +85,8 @@ aea add-key ethereum eth_private_key.txt
 |        country: UK                |        country: UK               |
 |        city: Cambridge            |        city: Cambridge           |
 |----------------------------------------------------------------------| 
-```
-
-``` yaml
+``` 
+``` yaml 
 |----------------------------------------------------------------------|
 |         FETCHAI                   |           ETHEREUM               |
 |-----------------------------------|----------------------------------|
@@ -80,30 +105,4 @@ aea add-key ethereum eth_private_key.txt
 |        constraint_type: '=='      |        constraint_type: '=='     |
 |ledgers: ['fetchai']               |ledgers: ['ethereum']             |
 |----------------------------------------------------------------------| 
-```
-
-``` bash
-aea generate-wealth fetchai
-```
-
-``` bash
-aea generate-wealth ethereum
-```
-
-```bash
-addr: ${OEF_ADDR: 127.0.0.1}
-```
-
-```bash 
-aea add connection fetchai/oef:0.1.0
-aea install
-aea run --connections fetchai/oef:0.1.0
-```
-
-```bash 
-cd ..
-aea delete my_seller_aea
-aea delete my_buyer_aea
-```
-
-
+```
