+++ tests/test_docs/test_bash_yaml/md_files/bash-gym-example.md
@@ -1,7 +1,6 @@
-``` bash
+``` bash 
 pip install numpy gym
-```
-
-``` bash
+``` 
+``` bash 
 python examples/gym_ex/train.py
-```
+```
