+++ tests/test_docs/test_bash_yaml/md_files/bash-ml-skills.md
@@ -1,86 +1,69 @@
-``` bash
+``` bash 
 python scripts/oef/launch.py -c ./scripts/oef/launch_config.json
-```
-
-``` bash
-aea create my_weather_station
-```
-
-``` bash
-cd my_weather_station
+``` 
+``` bash 
+aea create ml_data_provider
+cd ml_data_provider
 aea add connection fetchai/oef:0.1.0
-aea add skill fetchai/weather_station:0.1.0
+aea add skill fetchai/ml_data_provider:0.1.0
 aea install
-```
-
-``` bash
-aea config set vendor.fetchai.skills.weather_station.models.strategy.args.is_ledger_tx False --type bool
-```
-
-``` bash
+``` 
+``` bash 
+aea fetch fetchai/ml_data_provider:0.1.0
+cd ml_data_provider
+``` 
+``` bash 
+aea install
+``` 
+``` bash 
 aea run --connections fetchai/oef:0.1.0
-```
-
-``` bash
-aea create my_weather_client
-```
-
-``` bash
-cd my_weather_client
+``` 
+``` bash 
+aea create ml_model_trainer
+cd ml_model_trainer
 aea add connection fetchai/oef:0.1.0
-aea add skill fetchai/weather_client:0.1.0
+aea add skill fetchai/ml_train:0.1.0
 aea install
-```
-
-``` bash
-aea config set vendor.fetchai.skills.weather_client.models.strategy.args.is_ledger_tx False --type bool
-```
-
-``` bash
+``` 
+``` bash 
+aea fetch fetchai/ml_model_trainer:0.1.0
+cd ml_model_trainer
+``` 
+``` bash 
+aea install
+``` 
+``` bash 
 aea run --connections fetchai/oef:0.1.0
-```
-
-``` bash
-cd ..
-aea delete my_weather_station
-aea delete my_weather_client
-```
-
-``` bash
-aea create my_weather_station
-cd my_weather_station
+``` 
+``` bash 
+aea create ml_data_provider
+cd ml_data_provider
 aea add connection fetchai/oef:0.1.0
-aea add skill fetchai/weather_station:0.1.0
+aea add skill fetchai/ml_data_provider:0.1.0
 aea install
-```
-
-``` bash
-aea create my_weather_client
-cd my_weather_client
+``` 
+``` bash 
+aea create ml_model_trainer
+cd ml_model_trainer
 aea add connection fetchai/oef:0.1.0
-aea add skill fetchai/weather_client:0.1.0
+aea add skill fetchai/ml_train:0.1.0
 aea install
-```
-
-```bash
+``` 
+``` bash 
 aea generate-key fetchai
 aea add-key fetchai fet_private_key.txt
-```
-
-```bash
+``` 
+``` bash 
 aea generate-key ethereum
 aea add-key ethereum eth_private_key.txt
-```
-
-``` bash
+``` 
+``` bash 
 aea generate-wealth fetchai
-```
-
-``` bash
+``` 
+``` bash 
 aea generate-wealth ethereum
-```
-
-``` yaml
+``` 
+``` bash 
 |----------------------------------------------------------------------|
 |         FETCHAI                   |           ETHEREUM               |
 |-----------------------------------|----------------------------------|
@@ -88,17 +71,21 @@ aea generate-wealth ethereum
 |  strategy:                        |  strategy:                       |
 |     class_name: Strategy          |     class_name: Strategy         |
 |    args:                          |    args:                         |
-|      date_one: "1/10/2019"        |      date_one: "1/10/2019"       |
-|      date_two: "1/12/2019"        |      date_two: "1/12/2019"       |
-|      price_per_row: 1             |      price_per_row: 1            |
+|      price_per_data_batch: 100    |      price_per_data_batch: 100   |
+|      batch_size: 2                |      batch_size: 2               |
 |      seller_tx_fee: 0             |      seller_tx_fee: 0            |
+|      buyer_tx_fee: 10             |      buyer_tx_fee: 10            |
+|      dataset_id: 'fmnist'         |      dataset_id: 'fmnist'        |
 |      currency_id: 'FET'           |      currency_id: 'ETH'          |
 |      ledger_id: 'fetchai'         |      ledger_id: 'ethereum'       |
 |      is_ledger_tx: True           |      is_ledger_tx: True          |
 |----------------------------------------------------------------------| 
-```
-
-``` yaml
+``` 
+``` bash 
+aea config set vendor.fetchai.skills.ml_data_provider.models.strategy.args.currency_id ETH
+aea config set vendor.fetchai.skills.ml_data_provider.models.strategy.args.ledger_id ethereum
+``` 
+``` bash 
 |----------------------------------------------------------------------|
 |         FETCHAI                   |           ETHEREUM               |
 |-----------------------------------|----------------------------------|
@@ -106,28 +93,36 @@ aea generate-wealth ethereum
 |  strategy:                        |  strategy:                       |
 |     class_name: Strategy          |     class_name: Strategy         |
 |    args:                          |    args:                         |
-|      max_price: 4                 |      max_price: 40               |
-|      max_buyer_tx_fee: 1          |      max_buyer_tx_fee: 200000    |
+|      dataset_id: 'fmnist'         |      dataset_id: 'fmnist'        |
+|      max_unit_price: 70           |      max_unit_price: 70          |
+|      max_buyer_tx_fee: 20         |      max_buyer_tx_fee: 20        |
 |      currency_id: 'FET'           |      currency_id: 'ETH'          |
 |      ledger_id: 'fetchai'         |      ledger_id: 'ethereum'       |
 |      is_ledger_tx: True           |      is_ledger_tx: True          |
-|ledgers: ['fetchai']               |ledgers: ['ethereum']             |
 |----------------------------------------------------------------------| 
-```
-
-``` bash
-aea config set vendor.fetchai.skills.weather_client.models.strategy.args.max_buyer_tx_fee 10000 --type int
-aea config set vendor.fetchai.skills.weather_client.models.strategy.args.currency_id ETH
-aea config set vendor.fetchai.skills.weather_client.models.strategy.args.ledger_id ethereum
-aea config set vendor.fetchai.skills.weather_client.models.strategy.args.is_ledger_tx True --type bool
-```
-
-``` bash
+``` 
+``` bash 
+aea config set vendor.fetchai.skills.ml_train.models.strategy.args.max_buyer_tx_fee 10000 --type int
+aea config set vendor.fetchai.skills.ml_train.models.strategy.args.currency_id ETH
+aea config set vendor.fetchai.skills.ml_train.models.strategy.args.ledger_id ethereum
+``` 
+``` bash 
 aea run --connections fetchai/oef:0.1.0
-```
-
-``` bash
+``` 
+``` bash 
 cd ..
-aea delete my_weather_station
-aea delete my_weather_client
-```
+aea delete ml_data_provider
+aea delete ml_model_trainer
+``` 
+``` yaml 
+ledger_apis:
+  fetchai:
+    network: testnet
+``` 
+``` yaml 
+ledger_apis:
+  ethereum:
+    address: https://ropsten.infura.io/v3/f00f7b3ba0e848ddbdc8941c527447fe
+    chain_id: 3
+    gas_price: 50
+```
