+++ tests/test_docs/test_bash_yaml/md_files/bash-thermometer-skills-step-by-step.md
